migrate(routing): extract the routing block from agent_config.c into the routing module (Class A #3)
Third Class A migration, and the first file *split*. The self-contained routing block of
src/server/agent_config.c — role dispatch (agent_route/at_tier/with_caps), capability/tier selection,
delegate pick, availability, route-block reasons, and the route health/policy filters (~503 lines) —
is extracted into src/modules/routing/routing.c and the descriptor latches. 21/26 latched; five Class
A modules remain. The config/auth half of agent_config.c stays in the server.

Why the split is link-clean (self-reviewed; roundtable unavailable, operator-authorized self-review):
- the block's statics (agent_supports_role, agent_command_on_path, agent_pick_random,
  delegate_role_rand, agent_satisfies_required_caps, agent_route_with_caps_inner, the g_route_* filter
  pointers) are each used only within the block — verified none is referenced outside it — so they
  move and stay static;
- no config/auth function calls the routing functions (no config->routing coupling);
- the block's three calls to functions that stay (agent_load_config, agent_has_role,
  agent_has_resolvable_credentials) go through the shared src/headers/agent_config.h — an established
  pattern, since the delegates module's delegate_routing.c already calls those header predicates.

The routing contract stays in agent_config.h and routing.c implements it while the server implements
the config/auth half (the memory/DB2 arrangement), so routing has no module-private header
(private_headers empty; module root holds only routing.c, latch exact). tests declares
test_agent_caps.c (capability routing). routing.c is added to every source/object/test-link list that
carries agent_config.c/.o. delegate_routing.c stays in delegates.

Verified: all module/descriptor checks + the descriptor suite pass (empty-root single-file test
repointed to control-web; routing gains its latched mutation coverage); aimee and aimee-server build
and link routing.o with the config half. Full Make + CMake + routing-test .253 validation to follow.
The linker enforces the partition.
