+++ src/modules/routing/module.yaml
@@ -10,5 +10,16 @@
   ],
   "runtime_toggle": {
     "supported": false
-  }
+  },
+  "ownership_complete": true,
+  "sources": [
+    "src/modules/routing/routing.c"
+  ],
+  "private_headers": [],
+  "tests": [
+    "src/tests/test_agent_caps.c"
+  ],
+  "docs": [
+    "docs/modules/routing.md"
+  ]
 }
